fix drive_simulation() so it can be called multiple times
MPI_Init can only be called once and that's currently done in this
function. For now, I make the call static so it only happens
once. There's probably a saner way to do this but it's a start. For
instance, if argc/argv change from one call to the next then MPI_Init
may not be initialized as expected after the first call.