Instantiation of the Simulation/Physics design
In order to allow multiple meshes to be used simultaneously with different
physics on each mesh, etc, we introduce a Simulation that will manage
the overall procedure and a Physics that will have its own mesh,
equations, and BCs. There will only be one simulation allowed, but
there can be an arbitrary number of physics. A physics will manage the
creation of all the necessary fields in the stk mesh and the i/o for
each Analysis_Model. The Analysis model holds the input and output files
as well as everything necessary to conver the stk mesh to exodusII.

Signed-off-by: Dan Turner <danielzturner@gmail.com>